AI_HR_Hiring_Assistant_Streamlit/requirements.txt
streamlit
pandas
numpy
scikit-learn
sentence-transformers
torch
pdfminer.six
python-docx